==> enhancer_driven/first_round/probe_design/02.2_Prepare-LD-Posfile-eQTL-CredSNPsOnly.Rmd <==
---
title: "02.1c_Prepare-LD-Posfile-eQTL-CredSNPsOnly.Rmd"
author: "Jason Torres"
date: "March 1, 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Setup libraries and directories 

```{r}
"%&%" <- function(a,b) paste0(a,b)

library("dplyr")
library("data.table")
library("ggplot2")

serv.dir <- "/Users/jtorres/FUSE/"
rds.dir <- serv.dir %&% "projects/wtsa/enhancer_driven/probe_design/rds/"

save.dir <- serv.dir %&% "projects/wtsa/enhancer_driven/probe_design/ld/"
dir.create(save.dir)
save.dir <- save.dir %&% "positions/"
dir.create(save.dir)

cred.limit <- 100 # snps 
credt2d.df <- readRDS(rds.dir%&%"credComb.df.RDS")

vec <- as.integer(table(credt2d.df$SEGNUMBER))
names(vec) <- names(table(credt2d.df$SEGNUMBER))

keep.ids <- names(vec[vec <=100])
credt2d.df <- filter(credt2d.df,SEGNUMBER %in% keep.ids)
saveRDS(credt2d.df,rds.dir%&%"credComb.100less.df.RDS")

```


Create Position files for credible set with positons for credible SNPs and significant eQTLs 

```{r}


eqtl.df <- readRDS(rds.dir%&%"eqtl.fdr01.df.RDS")

window<-1e6

pb <- txtProgressBar(min = 0, max = length(unique(credt2d.df$SEGNUMBER)), 
                     initial = 0, style = 3)
for (i in 1:length(unique(credt2d.df$SEGNUMBER))){
  seg <- unique(credt2d.df$SEGNUMBER)[i]
  setTxtProgressBar(pb,i)
  sub.df <- filter(credt2d.df,SEGNUMBER==seg)
  mn <- min(sub.df$POS); mx <- max(sub.df$POS); 
  chrom<-sub.df$CHR[1]; locus <- sub.df$LOCUS[1]
  study<- sub.df$STUDY[1]
  locus <- gsub("/","-",locus)
  locus <- gsub(" ","-",locus)
  locus <- gsub("(","",locus,fixed=TRUE); locus <- gsub(")","",locus,fixed=TRUE)
  mymin <- mn-window; mymax<-mx+window
  fname <- chrom%&%"_"%&%mymin%&%"-"%&%mymax%&%"_"%&%locus%&%"_" %&%
    study %&%  ".positions"
  eqtl.sub <- filter(eqtl.df,CHR==chrom,POS>=mymin,POS<=mymax)
  pos <- unique(c(sub.df$POS,eqtl.sub$POS))
  chr <- rep(gsub("chr","",chrom),length(pos))
  df <- data.frame(chr,pos); names(df) <- c("#chr","pos")
  write.table(x=df,file=save.dir%&%fname,quote=F,sep="\t",row.names=F,col.names=T)
}

```
