==> enhancer_driven/first_round/probe_design/04.2_TrackPlot.R <==

"%&%" <- function(a,b) paste0(a,b) 
library("data.table")
library("dplyr")
library("GenomicRanges")

serv.dir <- "/Users/jtorres/FUSE/"
#serv.dir <- "/well/got2d/jason/"
rds.dir <- serv.dir %&% "projects/wtsa/enhancer_driven/probe_design/rds/"


credt2d.df <- readRDS(rds.dir%&%"credt2d.df.RDS")
eqtl.df <- readRDS(rds.dir%&%"eqtl.df.RDS")
atac.df <- readRDS(rds.dir%&%"atac.df.RDS")


eqt2d.dir <- serv.dir %&% "reference/islet/eqtls/from_anthony/t2d_credible_regions/"
eqt2d.df <- fread("cat " %&% eqt2d.dir %&% 
                    "cis_eQTL_genelevel-t2d-credible.txt.gz" %&% " | zmore")


dpn.file <- serv.dir %&% "reference/DpnII/hg19_DpnII-sites.bed.gz" #  0-based bed file 
dpn.df <- fread("cat " %&% dpn.file %&% " | zmore")
# adjust to 1-based scheme 
dpn.df$V2 <- dpn.df$V2 + 1 
dpn.df$V3 <- dpn.df$V3 + 1 


gread <- function(bed.df){
  # names must be c(chr, start, end, id, score, strand), where available 
  if(length(bed.df)==3){
    gr <- with(bed.df, GRanges(chr, IRanges(start, end)))
  } else if (length(bed.df)==4){
    gr <- with(bed.df, GRanges(chr, IRanges(start, end), id=id))
  } else if (length(bed.df)==5){
    gr <- with(bed.df, GRanges(chr, IRanges(start, end), id=id, score=score))
  } else if (length(bed.df)>=6){
    bed.df$strand <- gsub(pattern="[^+-]+",replacement='*', bed.df$strand)
    gr <- with(bed.df, GRanges(chr, IRanges(start, end), id=id, score=score, strand=strand))
  }
  return(gr)
}

get_overlaps <- function(a,b){
  # a and b are genomic ranges
  vec <- a %over% b
  ##print(sum(vec) %&% " / " %&% length(vec))
  return(vec)
}


atac_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS); span <- mx - mn
  chrom <- df$CHR[1];mymin <- (mn-fac*span);mymax <- (mx+fac*span)
  a1sub <- filter(atac.df,chr==chrom,start>=mymin,end<=mymax,id=="Oxford") %>% 
    dplyr::select(one_of("chr","start","end"))
  a1.g <- gread(a1sub[,1:3])
  a2sub <- filter(atac.df,chr==chrom,start>=mymin,end<=mymax,id=="Parker") %>% 
    dplyr::select(one_of("chr","start","end"))
  a2.g <- gread(a2sub[,1:3])
  #if (length(a1.g)==0){
  #  a1.g <- GRanges(seqnames=chrom,IRanges(start=0,end=0),id="null")
  #}
  #if (length(a2.g)==0){
  #  a2.g <- GRanges(seqnames=chrom,IRanges(start=0,end=0),id="null")
  #}
  if (length(a1.g)>0 & length(a2.g)>0){
    plt <- ggplot(a2.g) + geom_segment(color="firebrick3",size=30,alpha=0.9) +
      geom_segment(a1.g,color="yellow2",size=25,alpha=0.9) + ylab("") +
      theme_bw() + theme(panel.grid = element_blank())
  } else if (length(a1.g)>0){
    plt <- ggplot(a1.g) + geom_segment(color="yellow2",size=25,alpha=0.9) +
      theme_bw() + theme(panel.grid = element_blank())
  } else if (length(a2.g)>0){
    plt <- ggplot(a2.g) + geom_segment(color="firebrick3",size=30,alpha=0.9) +
      theme_bw() + theme(panel.grid = element_blank())
  } else {
    dframe <- data.frame(start=mymin,end=mymax)
    plt <- ggplot(data=dframe) + geom_rect(data=dframe,aes(xmin=start,xmax=end),
                                           ymin=0,ymax=0) + theme_bw()
  }
  return(plt)
}



dpnII_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS); span <- mx - mn
  chrom <- df$CHR[1];mymin <- (mn-fac*span);mymax <- (mx+fac*span)
  dpnsub <- filter(dpn.df,V1==chrom,V2>=mymin,V3<=mymax)
  if (dim(dpnsub)[1]==0){
    dpnsub <- data.frame(V1=chrom,V2=mymin,V3=mymax)
    plt <-  ggplot(data=dpnsub)
    plt <- plt + geom_vline(aes(xintercept=V2), size=0,color="white") + 
      geom_vline(aes(xintercept=V3),size=0,color="white")
  } else{
    plt <-  ggplot(data=dpnsub)
    plt <- plt + geom_vline(aes(xintercept=V2), size=0.1) + 
      geom_vline(aes(xintercept=V3),size=0.1) 
  }
  plt <- plt + ylim(c(0,1)) + 
    theme_bw() + theme(axis.text.y=element_blank(),
                       axis.ticks = element_blank(),
                       panel.grid = element_blank())  
  
  return(plt)
}


credld.dir <- serv.dir %&% "reference/gwas/diagram_1Kgenomes/" %&% 
  "credible_sets/ld_1Kgenomes/ld_proxies/credible_snps_only/"

credld_plot <- function(segnum,fac=10){
  df <- filter(credt2d.df,SEGNUMBER==segnum) %>% arrange(desc(PPA))
  mx <- max(df$POS); mn <- min(df$POS); span <- mx - mn; loc <- df$LOCUS[1]
  chrom <- df$CHR[1];mymin <- (mn-fac*span);mymax <- (mx+fac*span)  
  loc <- gsub("/","-",loc)
  f <- chrom%&%"_"%&%mn%&%"-"%&%mx%&%"_"%&%loc%&%".hap.ld"
  temp.df <- fread(credld.dir %&% f)
  temp.df <- temp.df[,1:7]
  names(temp.df) <- c("CHR","POS1",	"POS2","N_CHR","R2","D","Dprime")
  maxpos <- df$POS[1] # by PPA 
  LeadR2 <- as.numeric(sapply(1:length(df$POS),function(i){
    pos <- df$POS[i]
    if (pos==maxpos){r2<-1}
    else {r2<-filter(temp.df,(POS1==maxpos & POS2==pos) | 
                       (POS1==pos & POS2==maxpos))$R2}
    return(r2)
  }))
  Color <- as.character(sapply(1:length(LeadR2),function(i){
    r2 <- LeadR2[i]
    col <- ifelse(r2>=0.80,"red", 
                  ifelse(r2<0.8 & r2>=0.6, "yellow",
                         ifelse(r2<0.6 & r2>=0.4, "green",
                                ifelse(r2<0.4 & r2>=0.2, "dodgerblue1",
                                       ifelse(r2<0.2 & r2>=0.0, "purple4",
                                              "purple4")))))
  }))
  df <- cbind(df,LeadR2,Color)
  plt <- ggplot(data=df,aes(x=POS,y=PPA)) + 
    geom_point(shape=21,color="black",size=2,fill=df$Color) + 
    scale_fill_continuous(low="yellow",
                          high="firebrick1",limits=c(0,1),
                          breaks=seq(0,1,0.2)) + 
    scale_y_continuous(limits=c(0,1),breaks=seq(0,1,0.1)) + 
    theme_bw() + 
    theme(legend.position="none",
          panel.grid.minor=element_blank(),
          axis.title.y=element_text(size=5),
          axis.text.y=element_text(size=4),
          panel.grid=element_line(size=0.1)) +
    scale_y_continuous(breaks=seq(0,1,0.2),limits=c(0,1)) + 
    coord_cartesian(ylim=c(0,1),expand=TRUE)
  return(plt)
}



fgwas_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS); span <- mx - mn
  chrom <- df$CHR[1];mymin <- (mn-fac*span);mymax <- (mx+fac*span)
  plt <- ggplot(df,aes(x=POS)) + 
    geom_point(aes(y=PPA),shape=21,
               color="gray",fill="gray") + 
    geom_point(aes(y=PPA.fgwas,fill=(change>0)),
               shape=23,color="black") +   
    scale_fill_manual(values=c("dodgerblue2","firebrick2")) + 
    xlab("Position on Chromosome " %&% chrom) + 
    theme_bw() +  
    theme(legend.position="none",
          panel.grid.minor=element_blank(),
          panel.grid=element_line(size=0.1),
          axis.title.y=element_text(size=5),
          axis.text.y=element_text(size=4)) +
    scale_y_continuous(breaks=seq(0,1,0.2),limits=c(0,1)) + 
    coord_cartesian(ylim=c(0,1),expand=TRUE)
  return(plt)
}


library(ggbio)
library(Homo.sapiens)
gene_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS)
  chrom <- df$CHR[1]; span <- mx - mn
  gr <- GRanges(chrom,IRanges(start=mn-1e6,end=mx+1e6))
  aplt <- autoplot(Homo.sapiens, which = gr, label.color = "black",#
                   color = "brown",
                   #fill = "brown") + #,stat="reduce"
                   fill = "brown",stat="reduce") + 
    xlim(c(mn,mx)) + 
    theme_alignment(grid = FALSE,border = FALSE) 
  return(aplt)
}

islet_eqtl_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS)
  chrom <- df$CHR[1]; loc <- df$LOCUS[1]; span <- mx - mn
  mymin <- (mn-fac*span); mymax <- (mx+fac*span)
  sub.df <- filter(eqt2d.df,chr==chrom,pos>=mymin) %>% filter(pos<=mymax)
  e.sub <- filter(eqtl.df, CHR==chrom, POS>=mymin) %>% filter(POS<=mymax)
  if (dim(sub.df)[1]==0){
    sub.df <- data.frame(snp=NA,chr=NA,pos=0,gene=NA,beta=NA,tstat=NA,
                         pval=1,qval=NA)
  }
  #if (dim(e.sub)[1]==0){
  #  e.sub <- data.frame(RSID=NA,ENSID=NA,Beta=NA,`T`=NA,P=1,Q=NA,CHR=NA,
  #                      POS=0,GENE=NA,SNPID=NA)
  #}
  plt <- ggplot(data=sub.df, aes(x=pos,y=-log(pval,base=10))) +  
    geom_point(shape=21,color="grey", size=2,
               fill="grey",alpha=0.80) + 
    ylab(expression(paste("-log"[10],"(p-value)"))) + theme_bw() + 
    theme(panel.grid = element_blank(),
          axis.title.y=element_text(size=5),
          axis.text.y=element_text(size=4)) +  xlim(c(mn,mx)) 
  if (dim(e.sub)[1]>0){
    plt <- plt + geom_point(data=e.sub, aes(x=POS,y=-log(P,base=10)),
                            color="black", shape=21,stroke=1.1,
                            fill="green1",size=2) 
  }
  return(plt)
}


prob.dir <- serv.dir %&% "projects/wtsa/enhancer_driven/" %&% "probe_design/from_Martijn/DpnII_120bp/"
probe.df <- fread(prob.dir %&% "lessThan100_all_success.txt")
chrom <- as.character(sapply(1:dim(probe.df)[1],function(i){
  frag <- probe.df$`Fragment Coordinates`[i]
  c <- strsplit(x=frag,split=":")[[1]][1]
  return(c)
}))
frag.start <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  frag <- probe.df$`Fragment Coordinates`[i]
  span <- strsplit(x=frag,split=":")[[1]][2]
  fs <- as.integer(strsplit(x=span,split="-")[[1]][1])
  return(fs)
}))
frag.end <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  frag <- probe.df$`Fragment Coordinates`[i]
  span <- strsplit(x=frag,split=":")[[1]][2]
  fe <- as.integer(strsplit(x=span,split="-")[[1]][2])
  return(fe)
}))
left.start <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  span <- probe.df$`Left Oligo`[i]
  ls <- as.integer(strsplit(x=span,split="-")[[1]][1])
  return(ls)
}))
left.end <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  span <- probe.df$`Left Oligo`[i]
  le <- as.integer(strsplit(x=span,split="-")[[1]][2])
  return(le)
}))
right.start <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  span <- probe.df$`Right Oligo`[i]
  rs <- as.integer(strsplit(x=span,split="-")[[1]][1])
  return(rs)
}))
right.end <- as.integer(sapply(1:dim(probe.df)[1],function(i){
  span <- probe.df$`Right Oligo`[i]
  re <- as.integer(strsplit(x=span,split="-")[[1]][2])
  return(re)
}))
probe.df <- cbind(probe.df,chrom,frag.start,frag.end,
                  left.start,left.end,right.start,right.end)



prob_plot <- function(segnum,fac=0.1){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS)
  mychrom <- df$CHR[1]; loc <- df$LOCUS[1]; span <- mx - mn
  mymin <- (mn-fac*span); mymax <- (mx+fac*span)
  sub <- filter(probe.df,chrom==mychrom)
  if (dim(sub)[1]==0){
    sub <- data.frame(a=NA,b=NA,c=NA,d=NA,e=NA,f=NA,g=NA,h=NA,i=NA,j=NA,k=NA,l=NA)
    names(sub) <- names(probe.df)
  }
  plt <- ggplot(data=sub) + 
    geom_rect(sub, fill="black", aes(xmin=frag.start,xmax=frag.end),
              ymin=0.4, ymax=0.6) + 
    geom_rect(sub, fill="dodgerblue2", aes(xmin=left.start,xmax=left.end),
              ymin=0.6, ymax=0.7) +
    geom_rect(sub, fill="dodgerblue2", aes(xmin=right.start,xmax=right.end),
              ymin=0.3, ymax=0.4) +
    coord_cartesian(xlim=c(mymin,mymax)) + theme_bw() + 
    theme(panel.grid = element_blank())
  return(plt)
}

imr90 <- fread(serv.dir%&%
                 "reference/tads/IMR90/combined/total.combined.domain")
hesc <- fread(serv.dir%&%
                "reference/tads/hESC/combined/total.combined.domain")
cell <- c(rep("IMR90",dim(imr90)[1]),rep("hESC",dim(hesc)[1]))
tad.df <- as.data.frame(cbind(rbind(imr90,hesc),cell))
names(tad.df) <- c("chr","start","end","id")
tad.df$start <- tad.df$start + 1
tad.df$end <- tad.df$end + 1

tad.gr <- gread(tad.df)

tad_plot <- function(segnum,fac=0.1){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS)
  mychrom <- df$CHR[1]; loc <- df$LOCUS[1]; span <- mx - mn
  mymin <- (mn-fac*span); mymax <- (mx+fac*span) 
  gr <- GRanges(mychrom,IRanges(mymin,mymax))
  df <- as.data.frame(tad.gr[tad.gr %over% gr])
  yhigh <- as.numeric(as.factor(df$id))
  ylow <- as.numeric(as.factor(df$id)) - 1 
  df <- cbind(df,ylow,yhigh)
  plt <- ggplot(data=df,aes(xmin=start,xmax=end,
                            ymin=ylow,ymax=yhigh,fill=id)) + 
    geom_rect(color="black",alpha=0.3) + 
    scale_fill_manual(values=c("gold","dodgerblue4")) + 
    theme_bw() + 
    theme(panel.grid=element_blank(),
          axis.text.y=element_blank(),
          axis.ticks.y=element_blank(),
          legend.position="none") + 
    coord_cartesian(xlim=c(mymin,mymax))
  return(plt)
}


track_plot <- function(segnum,fac=0.10){
  df <- filter(credt2d.df,SEGNUMBER==segnum)
  mx <- max(df$POS); mn <- min(df$POS); span <- mx - mn
  mymin <- (mn-fac*span); mymax <- (mx+fac*span)
  loc <- df$LOCUS[1]
  atac_plt <- atac_plot(segnum,fac)
  dpn_plt <- dpnII_plot(segnum,fac)
  fgwas_plt <- fgwas_plot(segnum,fac)
  credld_plt <- credld_plot(segnum,fac)
  eqtl_plt <- islet_eqtl_plot(segnum,fac)
  gene_plt <- gene_plot(segnum,fac)
  probe_plt <- prob_plot(segnum,fac)
  tad_plt <- tad_plot(segnum,fac)
  
  tracks(`Islet eQTL`=eqtl_plt,
         `ATAC \nEndoC`=atac_plt,
         `Credible Set \n(Islet fGWAS)`=fgwas_plt,
         `Credible Set \n(LD)`=credld_plt,
         `Oligos`=probe_plt,
         `DpnII`=dpn_plt,
         `TAD`=tad_plt,
         `Genes`=gene_plt,
         heights=c(2,0.5,2,2,0.5,0.5,1,1),title=loc %&% "\nInterval: " %&% 
           
           round(span/1000,digits=2) %&%" kb",
         label.text.cex=0.6,main.height=2) +
    scale_x_sequnit("Mb") + 
    theme(axis.text.x=element_text(size=5)) 
}
