==> enhancer_driven/first_round/probe_design/06.1_PrioritizeLoci-T2D.Rmd <==
---
title: "06.1_PrioritizeLoci-T2D.Rmd"
author: "Jason Torres"
date: "February 17, 2017"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}


"%&%" <- function(a,b) paste0(a,b) 
library("data.table")
library("dplyr")

serv.dir <- "/Users/jtorres/FUSE/"
work.dir <- serv.dir %&% "projects/wtsa/enhancer_driven/probe_design/"
rds.dir <- work.dir %&% "rds/"
profile.dir <- work.dir %&% "profile-snps/"

source(serv.dir %&% "projects/wtsa/enhancer_driven/probe_design/" %&%
         "04.1b_VizTracks-T2D.R")

save.dir <- work.dir %&% "profile-loci/"
dir.create(save.dir)

fig.dir <- work.dir %&% "figures/"

```



Read in prioritized data fram 

```{r}
prof.df <- fread(profile.dir%&%"profile_credt2d.txt")
```


Build loci summary data frame 


```{r}

cumm_99 <- function(vec){
  # vec is a arranged (descending) vector of PPAs 
  # returns the indeces needed to get cummulative PPA of 0.99 
  sum = 0 
  out.vec <- c()
  for (i in 1:length(vec)){
    val <- vec[i]
    sum <- sum+val
    if (sum<=0.99){
      out.vec<-append(out.vec,i)
    }
  }
  return(out.vec)
}

build_summary <- function(df=prof.df){
  out.df <- c()
  pb <- txtProgressBar(min=0,max=length(unique(df$SEGNUMBER)),style=3)
  for (i in 1:length(unique(df$SEGNUMBER))){
    setTxtProgressBar(pb,i)
    seg <- unique(df$SEGNUMBER)[i]
    temp <- filter(df,SEGNUMBER==seg) %>% arrange(desc(PPA.fgwas))
    chr <- temp$CHR[1]; locus <- temp$LOCUS[1]; segnum <- temp$SEGNUMBER[1]
    top.endo.atac <- temp$endo.atac[1] 
    top.num.eqtls.ld80 <- temp$num.eqtls.ld80[1]
    top.egenes.ld80 <- temp$egenes.ld80[1]
    top.num.eqtls.ld20 <- temp$num.eqtls.ld20[1]
    top.egenes.ld20 <- temp$egenes.ld20[1]
    cred.num <- dim(temp)[1]
    if (is.na(temp$PPA.fgwas)){
      fgwas.cred.num <- length(cumm_99(temp$PPA))
    } else{
      fgwas.cred.num <- length(cumm_99(temp$PPA.fgwas))
    }
    if (fgwas.cred.num==0){
      fgwas.cred.num=1
    }
    stack.df <- data.frame(segnum,chr,locus,cred.num,fgwas.cred.num,
                           top.endo.atac,top.num.eqtls.ld80,top.egenes.ld80,
                           top.num.eqtls.ld20,top.egenes.ld20)
    out.df <- rbind(out.df,stack.df)
  }
  out.df$chr <- as.character(out.df$chr)
  out.df$locus <- as.character(out.df$locus)
  out.df$top.egenes.ld80 <- as.character(out.df$top.egenes.ld80)
  out.df$top.egenes.ld20 <- as.character(out.df$top.egenes.ld20)
  prop.reduced <- ifelse(out.df$fgwas.cred.num>=1 & out.df$cred.num!=1,
                         (1 - (out.df$fgwas.cred.num/out.df$cred.num)) ,NA)
  out.df <- cbind(out.df,prop.reduced)
  out.df <- arrange(out.df,desc(prop.reduced))
  out.df$segnum <- as.character(out.df$segnum)
  return(out.df)
}

sum.df <- build_summary()

```

# Designate Tiers 


```{r}

build_tiers <- function(){
  top.tier <- filter(sum.df,top.endo.atac==TRUE) %>% arrange(desc(top.num.eqtls.ld20))
  
  bottom.tier <- filter(sum.df,top.endo.atac==FALSE) %>% 
    arrange(desc(top.num.eqtls.ld20))
  
  tier1 <- filter(top.tier,top.num.eqtls.ld80>0,fgwas.cred.num<=10)
  tier2 <- filter(top.tier,top.num.eqtls.ld20>0 & top.num.eqtls.ld80==0,fgwas.cred.num<=10)
  tier3 <- filter(top.tier,top.num.eqtls.ld80>0,fgwas.cred.num>10)
  tier4 <- filter(top.tier,top.num.eqtls.ld20>0 & top.num.eqtls.ld80==0,fgwas.cred.num>10)
  tier5 <- filter(top.tier,top.num.eqtls.ld20==0,fgwas.cred.num<=10)
  tier6 <- filter(top.tier,top.num.eqtls.ld20==0,fgwas.cred.num>10)
  
  tier7 <- filter(bottom.tier,top.num.eqtls.ld80>0,fgwas.cred.num<=10)
  tier8 <- filter(bottom.tier,top.num.eqtls.ld20>0 & top.num.eqtls.ld80==0,fgwas.cred.num<=10)

  tier9 <- filter(bottom.tier,top.num.eqtls.ld80>0,fgwas.cred.num>10)
  tier10 <- filter(bottom.tier,top.num.eqtls.ld20>0 & top.num.eqtls.ld80==0,fgwas.cred.num>10)

  tier11 <- filter(bottom.tier,top.num.eqtls.ld20==0,fgwas.cred.num<=10)
  tier12 <- filter(bottom.tier,top.num.eqtls.ld20==0,fgwas.cred.num>10)
  
  check = dim(tier1)[1] +  dim(tier2)[1]+ dim(tier3)[1]+ dim(tier4)[1]+ 
    dim(tier5)[1] + dim(tier6)[1]+ dim(tier7)[1] + 
    dim(tier8)[1] + dim(tier9)[1]+ dim(tier10)[1] + 
    dim(tier11)[1] + dim(tier12)[1]# sanity check   
  if (check==(dim(top.tier)[1]+dim(bottom.tier)[1])){
    tier <- c(rep(1,dim(tier1)[1]), rep(2,dim(tier2)[1]), rep(3,dim(tier3)[1]),
                rep(4,dim(tier4)[1]), rep(5,dim(tier5)[1]), rep(6,dim(tier6)[1]),
               rep(7,dim(tier7)[1]), rep(8,dim(tier8)[1]),rep(9,dim(tier9)[1]), rep(10,dim(tier10)[1]),
              rep(11,dim(tier11)[1]), rep(12,dim(tier12)[1]))
  }
  out.df <- rbind(tier1,tier2,tier3,tier4,tier5,
                  tier6,tier7,tier8,tier9,tier10,tier11,tier12)
  out.df <- cbind(out.df,tier)
  write.table(x=out.df,file=save.dir%&%"profile-t2d-loci.txt",
              row.names=F,sep="\t",quote=F)
  return(out.df)
}

tier.df <- build_tiers()

plt <- ggplot(data=filter(tier.df,cred.num<=100),
              aes(x=cred.num,y=fgwas.cred.num)) + 
  geom_point(shape=21,color="black",aes(fill=as.factor(tier))) + theme_bw() + 
  geom_abline(slope=1)



```

Save Plots 
Save Top Tiers 

```{r}

# For i=25, segnum="6D", use fac=10, manual adjustment 
# For i=26, segnum="35M", use fac=20, manual adjustment 

save_top <- function(){
  top.tier <- filter(sum.df,top.endo.atac==TRUE) %>% 
    arrange(desc(top.num.eqtls.ld20))
  for (i in 1:length(top.tier$segnum)){
#  for (i in 27:length(top.tier$segnum)){

    segnum <- top.tier$segnum[i]
    chr <- top.tier$chr[i]
    locus <- top.tier$locus[i]
    locus <- gsub("/","-",locus); locus <- gsub(" ","-",locus)
    locus <- gsub("(","",locus,fixed = TRUE)
    locus <- gsub(")","",locus,fixed = TRUE)
    savename <- fig.dir %&% segnum %&% "_" %&% chr %&% "_" %&% locus %&% ".png"
    print(paste("index: ",i, " segnum: ",segnum))
    plt <- viz_loc(segnum,fac=0.20)
    ggsave(savename,plot=plt,width=5,height = 8)
  }
}

save_top()

```

Save Bottom Tiers 

```{r}

# For i=51, segnum="18D", use fac=16, manual adjustment 
# For i=71, segnum="7T", use fac=50000, manual adjustment 


save_bottom <- function(){
  bot.tier <- filter(sum.df,top.endo.atac==FALSE) %>% 
    arrange(desc(top.num.eqtls.ld20))
#  for (i in 1:length(bot.tier$segnum)){
  for (i in 72:length(bot.tier$segnum)){

    segnum <- bot.tier$segnum[i]
    chr <- bot.tier$chr[i]
    locus <- bot.tier$locus[i]
    locus <- gsub("/","-",locus); locus <- gsub(" ","-",locus)
    locus <- gsub("(","",locus,fixed = TRUE)
    locus <- gsub(")","",locus,fixed = TRUE)
    savename <- fig.dir %&% segnum %&% "_" %&% chr %&% "_" %&% locus %&% ".png"
    print(paste("index: ",i, " segnum: ",segnum))
    plt <- viz_loc(segnum,fac=10)
    ggsave(savename,plot=plt,width=5,height = 8)
  }
}

save_bottom()



```



# Annotation stacked bar plots 

```{r}
annot.vec <- c("islet_state8","islet_state9","cds")
test <- filter(ptzd.df,STUDY=="DIAGRAM")
fgwas.df <- fread("cat " %&% serv.dir %&% 
                    "projects/t2d-integration/fgwas/credsets_diagram_1KG/fgwas_input/" %&% 
                    "fgwas_input27.txt.gz" %&%  " | zmore") %>% dplyr::select(one_of("SNPID",annot.vec))
#fgwas.df$SEGNUMBER <- paste0(fgwas.df$SEGNUMBER,"D")
test <- inner_join(test,fgwas.df,by="SNPID"); test <- test[!duplicated(test),]
out.df <- c()
for (i in 1:length(unique(test$SEGNUMBER))){
  seg = unique(test$SEGNUMBER)[i]
  loc = unique(test$LOCUS)[i] 
  sub <- filter(test,SEGNUMBER==seg)
  a1 <- sum(sub$islet_state8)
  a2 <- sum(sub$islet_state9)
  a3 <- sum(sub$cds)
  a4 <- dim(sub)[1] - (a1+a2+a3)
  myrow <- data.frame(seg,loc,a1,a2,a3,a4,stringsAsFactors = F)
  out.df <- rbind(out.df,myrow)
}

out2.df <- c()
for (i in 1:dim(out.df)[1]){
  seg <- out.df$seg[i]; loc <- out.df$loc[i]
  a1 <- out.df$a1[i]; a2 <- out.df$a2[i]
  a3 <- out.df$a3[i]; a4 <- out.df$a4[i]
  out2.df <- rbind(out2.df,data.frame(seg,loc,count=a1,annot="a1",stringsAsFactors = FALSE))
  out2.df <- rbind(out2.df,data.frame(seg,loc,count=a2,annot="a2",stringsAsFactors = FALSE))
  out2.df <- rbind(out2.df,data.frame(seg,loc,count=a3,annot="a3",stringsAsFactors = FALSE))
  out2.df <- rbind(out2.df,data.frame(seg,loc,count=a4,annot="a4",stringsAsFactors = FALSE))
}

ggplot(data=out2.df,aes(x=reorder(seg,count),y=count,fill=annot)) + geom_bar(stat="identity") + coord_flip() + scale_fill_brewer(palette="Accent")

```

parameter CI_lo estimate CI_hi
pi_region 0.0682322 0.0874277 0.109222
islet_state8_ln 1.4097 2.2029 2.87315
cds_ln 1.20363 2.48094 3.35561
islet_state5_ln -3.43803 -1.60919 -0.575664
islet_atac_ln 0.123574 0.933363 1.65657
islet_state2_ln -3.21468 -1.65021 -0.677724

drop-islet_state6+transcript+islet_state13+intron+exon+islet_state7+islet_state9+distance_tss.params
